+++ src/icalendar/caselessdict.py
@@ -2,6 +2,11 @@
 from icalendar.parser_tools import to_unicode
 from icalendar.parser_tools import data_encode
 
+try:
+    from collections import OrderedDict
+except ImportError:
+    from ordereddict import OrderedDict
+
 
 def canonsort_keys(keys, canonical_order=None):
     """Sorts leading keys according to canonical_order.  Keys not specified in
@@ -20,7 +25,7 @@ def canonsort_items(dict1, canonical_order=None):
             k in canonsort_keys(dict1.keys(), canonical_order)]
 
 
-class CaselessDict(dict):
+class CaselessDict(OrderedDict):
     """A dictionary that isn't case sensitive, and only uses strings as keys.
     Values retain their case.
     """
@@ -28,47 +33,47 @@ class CaselessDict(dict):
     def __init__(self, *args, **kwargs):
         """Set keys to upper for initial dict.
         """
-        dict.__init__(self, *args, **kwargs)
+        super(CaselessDict, self).__init__(*args, **kwargs)
         for key, value in self.items():
             key_upper = to_unicode(key).upper()
             if key != key_upper:
-                dict.__delitem__(self, key)
+                super(CaselessDict, self).__delitem__(key)
                 self[key_upper] = value
 
     def __getitem__(self, key):
         key = to_unicode(key)
-        return dict.__getitem__(self, key.upper())
+        return super(CaselessDict, self).__getitem__(key.upper())
 
     def __setitem__(self, key, value):
         key = to_unicode(key)
-        dict.__setitem__(self, key.upper(), value)
+        super(CaselessDict, self).__setitem__(key.upper(), value)
 
     def __delitem__(self, key):
         key = to_unicode(key)
-        dict.__delitem__(self, key.upper())
+        super(CaselessDict, self).__delitem__(key.upper())
 
     def __contains__(self, key):
         key = to_unicode(key)
-        return dict.__contains__(self, key.upper())
+        return super(CaselessDict, self).__contains__(key.upper())
 
     def get(self, key, default=None):
         key = to_unicode(key)
-        return dict.get(self, key.upper(), default)
+        return super(CaselessDict, self).get(key.upper(), default)
 
     def setdefault(self, key, value=None):
         key = to_unicode(key)
-        return dict.setdefault(self, key.upper(), value)
+        return super(CaselessDict, self).setdefault(key.upper(), value)
 
     def pop(self, key, default=None):
         key = to_unicode(key)
-        return dict.pop(self, key.upper(), default)
+        return super(CaselessDict, self).pop(key.upper(), default)
 
     def popitem(self):
-        return dict.popitem(self)
+        return super(CaselessDict, self).popitem()
 
     def has_key(self, key):
         key = to_unicode(key)
-        return dict.__contains__(self, key.upper())
+        return super(CaselessDict, self).__contains__(key.upper())
 
     def update(self, indict):
         # Multiple keys where key1.upper() == key2.upper() will be lost.
@@ -76,10 +81,13 @@ class CaselessDict(dict):
             self[key] = value
 
     def copy(self):
-        return CaselessDict(dict.copy(self))
+        return type(self)(super(CaselessDict, self).copy())
 
     def __repr__(self):
         return 'CaselessDict(%s)' % data_encode(self)
+
+    def __eq__(self, other):
+        return self is other or dict(self.items()) == dict(other.items())
 
     # A list of keys that must appear first in sorted_keys and sorted_items;
     # must be uppercase.