+++ src/icalendar/tests/test_unit_cal.py
@@ -275,6 +275,37 @@ class TestCalComponent(unittest.TestCase):
             self.assertEqual(component[property_name].dt.tzinfo,
                              None)
 
+    def test_cal_Component_to_ical_property_order(self):
+        Component = icalendar.cal.Component
+        component_str = [b'BEGIN:VEVENT',
+                         b'DTSTART:19970714T170000Z',
+                         b'DTEND:19970715T035959Z',
+                         b'SUMMARY:Bastille Day Party',
+                         b'END:VEVENT']
+        component = Component.from_ical(b'\r\n'.join(component_str))
+
+        sorted_str = component.to_ical().splitlines()
+        assert sorted_str != component_str
+        assert set(sorted_str) == set(component_str)
+
+        preserved_str = component.to_ical(sorted=False).splitlines()
+        assert preserved_str == component_str
+
+    def test_cal_Component_to_ical_parameter_order(self):
+        Component = icalendar.cal.Component
+        component_str = [b'BEGIN:VEVENT',
+                         b'X-FOOBAR;C=one;A=two;B=three:helloworld.',
+                         b'END:VEVENT']
+        component = Component.from_ical(b'\r\n'.join(component_str))
+
+        sorted_str = component.to_ical().splitlines()
+        assert sorted_str[0] == component_str[0]
+        assert sorted_str[1] == b'X-FOOBAR;A=two;B=three;C=one:helloworld.'
+        assert sorted_str[2] == component_str[2]
+
+        preserved_str = component.to_ical(sorted=False).splitlines()
+        assert preserved_str == component_str
+
 
 class TestCal(unittest.TestCase):